utils: Add generator script for native apps
The script generates other scripts that setup, configure, build and run
native applications.

The input of the script is the `Kraftfile` of the application and the
`config.yaml` file that specifies configuration such as memory, path to
filesystem (if any), requirements for networking.

Also add the test support files (`test/common.sh` and `test/test.sh`) used
for testing. The application will define a `single_test.sh` script.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>