new-design: Separate build variants from run variants
Until this point a variant consisted of a combination of build and run
parameters. However, this was redundant as the the build parameters (and
build) were part of multiple variants.

Separate build parameters from run parameters in `tester.yaml`. A
variant consists of a build variant (set of parameters) and multiple run
variants (an array of sets of parameters).

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>